# pydantic-ai/pydantic-ai-agent-etl/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ETL AI Agent API",
    "description": "Natural language interface for ETL operations",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ai/query": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Query Agent",
        "description": "Process a natural language query using the ETL agent with OpenAI.\n\nThis endpoint accepts natural language queries and processes them using\nOpenAI with function calling to execute appropriate ETL operations.",
        "operationId": "query_agent_api_v1_ai_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Natural language query for the ETL agent"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "Optional user ID for the request"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest",
        "description": "Request model for natural language queries."
      },
      "QueryResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the query was successful"
          },
          "output": {
            "type": "string",
            "title": "Output",
            "description": "The agent's response to the query"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if query failed"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Unique identifier for this request"
          }
        },
        "type": "object",
        "required": [
          "success",
          "output",
          "request_id"
        ],
        "title": "QueryResponse",
        "description": "Response model for query results."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}